Refactor PrometheusRequestFailedException and add comments
