+++ mico-core/src/main/java/io/github/ust/mico/core/exception/PrometheusRequestFailedException.java
@@ -19,19 +19,11 @@
 
 package io.github.ust.mico.core.exception;
 
-import org.springframework.http.HttpStatus;
-
-public class PrometheusRequestFailedException extends Exception{
+public class PrometheusRequestFailedException extends Exception {
 
     public static final long serialVersionUID = 381515548494367817L;
 
-    // TODO: Either remove those fields or actually use them
-    private final HttpStatus httpStatus;
-    private final String prometheusResponseStatus;
-
-    public PrometheusRequestFailedException(String message, HttpStatus httpStatus, String prometheusResponseStatus){
+    public PrometheusRequestFailedException(String message) {
         super(message);
-        this.httpStatus = httpStatus;
-        this.prometheusResponseStatus = prometheusResponseStatus;
     }
 }
